refactor: rename "sidebar" folder to "sideBar"
