+++ packages/web-app-files/src/components/SideBar/SideBar.vue
@@ -52,7 +52,7 @@ import {
   useStore
 } from 'web-pkg/src/composables'
 import { bus } from 'web-pkg/src/instance'
-import { SideBarEventTopics } from '../../composables/sidebar'
+import { SideBarEventTopics } from '../../composables/sideBar'
 
 export default defineComponent({
   components: { FileInfo, SpaceInfo, SideBar },
