+++ packages/web-app-files/tests/unit/composables/sideBar/useSideBar.spec.ts
@@ -1,5 +1,5 @@
 import { EventBus } from 'web-pkg/src/event'
-import { SideBarEventTopics, useSideBar } from '../../../../src/composables/sidebar'
+import { SideBarEventTopics, useSideBar } from '../../../../src/composables/sideBar'
 import { unref } from '@vue/composition-api'
 import { createComposableWrapper } from '../composables.setup'
